Switch from `type: module` to .mjs-file
The esbuild configuration file utilizes the top-level `await` feature and the `import` constructions.

To support them, the file should be treated as a JS-module. Previously, it was achieved by declaring the whole project as a module. I decided to move away from that approach in favour of declaring a module only the config file itself by using the `.mjs` extension.