Fix the declaration of Node.js environment vars
Previously, the declaration was an extension of the list of native Node.js environment variables. It didn’t work. The compiler was highlighting the error and suggesting installing `@types/node`.

This solution would be incorrect, as Figma’s sandbox doesn’t provide any of those Node.js variables. So, to solve this issue, the declaration was rewritten as a simple global `process` variable not connected to the Node.js environment.